+++ src/mock/event.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils';
+import { getRandomInteger } from '../utils/utils';
 
 export const DESTINATIONS = [
   'Chamonix',
