+++ src/view/site-filters/site-filters.tpl.js
@@ -1,6 +1,4 @@
-import { createElement } from '../render';
-
-const createSiteFiltersTemplate = () => (
+export const createSiteFiltersTemplate = () => (
   `<form class="trip-filters" action="#" method="get">
     <div class="trip-filters__filter">
       <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
@@ -20,23 +18,3 @@ const createSiteFiltersTemplate = () => (
     <button class="visually-hidden" type="submit">Accept filter</button>
   </form>`
 );
-
-export default class SiteFilterView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  get template() {
-    return createSiteFiltersTemplate();
-  }
-
-  removeELement() {
-    this.#element = null;
-  }
-}
