+++ src/view/site-list-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import { createElement } from '../utils/render';
 
 const createPointListTemplate = () => (
   `<ul class="trip-events__list">
