+++ src/view/site-sort-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import { createElement } from '../utils/render';
 
 const createSiteSortTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
